project_root/requirements.txt
torch>=2.0.0
datasets>=2.14.0
torchtext>=0.15.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
nltk>=3.8.0
pyyaml>=6.0
numpy>=1.24.0
